requirements.txt
# PDF Processing Pipeline Dependencies
pypdf==3.17.4
camelot-py==1.0.0
pandas==2.2.3
numpy>=1.22.0  # Required by pandas and camelot
opencv-python-headless>=4.5.0  # Required by camelot
tabulate>=0.8.9  # Required for markdown tables
pdfminer-six>=20200517  # Required by camelot
Unidecode==1.3.8
chardet>=4.0.0  # Required for encoding detection
